+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/tuning/LocalizationTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.tuning;
+package org.firstinspires.ftc.teamcode.autonomous.tuning;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
@@ -8,9 +8,9 @@ import com.acmerobotics.roadrunner.PoseVelocity2d;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.Drawing;
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
+import org.firstinspires.ftc.teamcode.autonomous.Drawing;
+import org.firstinspires.ftc.teamcode.autonomous.MecanumDrive;
+import org.firstinspires.ftc.teamcode.autonomous.TankDrive;
 
 public class LocalizationTest extends LinearOpMode {
     @Override
