+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/tuning/ManualFeedbackTuner.java
@@ -1,13 +1,13 @@
-package org.firstinspires.ftc.teamcode.tuning;
+package org.firstinspires.ftc.teamcode.autonomous.tuning;
 
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
-import org.firstinspires.ftc.teamcode.ThreeDeadWheelLocalizer;
-import org.firstinspires.ftc.teamcode.TwoDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.autonomous.MecanumDrive;
+import org.firstinspires.ftc.teamcode.autonomous.TankDrive;
+import org.firstinspires.ftc.teamcode.autonomous.ThreeDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.autonomous.TwoDeadWheelLocalizer;
 
 public final class ManualFeedbackTuner extends LinearOpMode {
     public static double DISTANCE = 64;
